add privacy builtin field, add owner info on workspace, add get tags from all workspaces, fixes
