+++ databox/client/src/components/Dialog/Workspace/InfoWorkspace.tsx
@@ -6,6 +6,7 @@ import KeyIcon from '@mui/icons-material/Key';
 import EventIcon from '@mui/icons-material/Event';
 import InfoRow from '../Info/InfoRow';
 import {useTranslation} from 'react-i18next';
+import PersonIcon from "@mui/icons-material/Person";
 
 type Props = DataTabProps<Workspace>;
 
@@ -21,6 +22,12 @@ export default function InfoWorkspace({data, onClose, minHeight}: Props) {
                     icon={<KeyIcon />}
                 />
                 <Divider />
+                <InfoRow
+                    label={t('workspace.info.owner', `Owner`)}
+                    value={data.owner?.username ?? data.owner?.id ?? '-'}
+                    copyValue={data.owner?.id}
+                    icon={<PersonIcon />}
+                />
                 <InfoRow
                     label={t('workspace.info.creation_date', `Creation date`)}
                     value={data.createdAt}
